Fix some trace problems
This change is in preparation for future, better trace handling.
Bit fields don't have a size_type - so they need a cast to int or uint in a
trace statement.
Pointers are always 4 bytes in PPE.
Pointers cannot be converted to a uint32_t (%x) without an explicit cast.
The printf/trace format %x is 4 bytes, even %016x is 4 bytes. What is
needed is %llx or %016llx otherise a uint64_t will get truncated.

RTC: 161851
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/31189
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>

+++ src/sbefw/sbeFifoMsgUtils.C
@@ -364,10 +364,11 @@ uint32_t sbeDsSendRespHdr(const sbeRespGenHdr_t &i_hdr,
 
         // If there is a SBE internal failure
         if((i_hdr.primaryStatus != SBE_PRI_OPERATION_SUCCESSFUL) ||\
-            (i_hdr.secondaryStatus != SBE_SEC_OPERATION_SUCCESSFUL))
+           (i_hdr.secondaryStatus != SBE_SEC_OPERATION_SUCCESSFUL))
         {
             SBE_ERROR( SBE_FUNC" primaryStatus:0x%08X secondaryStatus:0x%08X",
-                       i_hdr.primaryStatus, i_hdr.secondaryStatus);
+                       (uint32_t)i_hdr.primaryStatus,
+                       (uint32_t)i_hdr.secondaryStatus);
 
             //Add FFDC data as well.
             //Generate all the fields of FFDC package
