+++ src/sbefw/sbecmdsram.C
@@ -93,7 +93,9 @@ uint32_t sbeOccSramAccess_Wrap(const bool i_isGetFlag)
         }
 
         SBE_INFO("mode [0x%08X] addr[0x%08X] len[0x%08X]",
-             l_req.mode, l_req.addr, l_req.len);
+             (uint32_t)l_req.mode,
+             (uint32_t)l_req.addr,
+             (uint32_t)l_req.len);
 
         // Get the Proc Chip Target to be passed in to the procedure call
         Target<fapi2::TARGET_TYPE_PROC_CHIP> l_proc = plat_getChipTarget();
